add policy_rule_id to session event for easier debugging of policy manager rules (Comment #12644)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/hades/src@42998 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ src/com/untangle/node/application_control/ApplicationControlApp.java
@@ -389,7 +389,7 @@ public class ApplicationControlApp extends NodeBase
             List<ApplicationControlProtoRule> protoList = settings.getProtoRules();
 
             // look at every active session and apply the current proto rules
-            public boolean isMatch(Long policyId, short protocol, int clientIntf, int serverIntf, InetAddress clientAddr, InetAddress serverAddr, int clientPort, int serverPort, Map<String, Object> attachments)
+            public boolean isMatch(Integer policyId, short protocol, int clientIntf, int serverIntf, InetAddress clientAddr, InetAddress serverAddr, int clientPort, int serverPort, Map<String, Object> attachments)
             {
                 // find the application and if missing leave the session alone
                 String application = (String) attachments.get(NodeSession.KEY_APPLICATION_CONTROL_APPLICATION);
